Move set up from ngModule to services
