+++ src/app/app.module.ts
@@ -1,10 +1,8 @@
 import { NgModule, ErrorHandler } from '@angular/core';
-import { IonicApp, IonicModule, IonicErrorHandler } from 'ionic-angular';
-import { CloudSettings, CloudModule } from '@ionic/cloud-angular';
+import { IonicApp, IonicModule } from 'ionic-angular';
+import { CloudModule } from '@ionic/cloud-angular';
 import { Http } from '@angular/http';
-import { Storage } from '@ionic/storage';
-import { AuthHttp, AuthConfig } from 'angular2-jwt';
-import * as Raven from 'raven-js';
+import { AuthHttp } from 'angular2-jwt';
 import { MyApp } from './app.component';
 
 import { HomePage } from '../pages/home/home';
@@ -21,30 +19,8 @@ import { InventoryData } from '../providers/inventory-data';
 import { StockpileData } from '../providers/stockpile-data';
 import { UserData } from '../providers/user-data';
 
-let storage = new Storage();
-
-export function getAuthHttp(http) {
-  return new AuthHttp(new AuthConfig({
-    tokenGetter: (() => storage.get('id_token')),
-  }), http);
-}
-
-const cloudSettings: CloudSettings = {
-  'core': {
-    'app_id': 'APP_ID'
-  }
-};
-
-Raven
-  .config('https://a64af160cb8746709c9743f23f666b9b@sentry.io/150943')
-  .install();
-
-export class RavenErrorHandler extends IonicErrorHandler {
-  handleError(err: any): void {
-    super.handleError(err);
-    Raven.captureException(err.originalError || err);
-  }
-}
+import { RavenErrorHandler } from '../services/raven-error-handler';
+import { getAuthHttp, cloudSettings } from '../services/auth-http';
 
 @NgModule({
   declarations: [
